GasPrice struct in a separate file with additional functions
